Make LESS rendering reusable
Allow multiple root css files
