+++ src/styles/styles.11ty.js
@@ -5,6 +5,7 @@ const postcss = require('postcss');
 const autoprefixer = require('autoprefixer');
 const atImport = require('postcss-import');
 const cssNano = require('cssnano');
+const LessRenderer = require('./LessRenderer');
 
 const permalink = 'styles.css';
 const input = './styles.less';
@@ -13,25 +14,8 @@ const input = './styles.less';
  * This reads a single LESS file from `input` and outputs a single CSS file to `permalink`.
  * It is expected that the LESS file will have imports to other LESS files.
  */
-module.exports = class {
-  async data() {
-    const rawFilepath = path.join(__dirname, input);
-    return {
-      permalink,
-      rawFilepath,
-      rawLess: (await fs.readFileSync(rawFilepath)).toString(),
-    };
-  }
-
-  async render({ rawLess, rawFilepath, site }) {
-    const lessResult = await less.render(rawLess, { filename: rawFilepath });
-    const plugins = [atImport, autoprefixer];
-    if (site.environment !== 'development') {
-      plugins.push(cssNano);
-    }
-    const postcssResult = await postcss(plugins).process(lessResult.css, {
-      from: input,
-    });
-    return postcssResult.css;
+module.exports = class AllStyles extends LessRenderer {
+  constructor() {
+    super({ input, permalink });
   }
 };
